+++ src/promptflow-recording/promptflow/recording/local/test_utils.py
@@ -7,14 +7,14 @@ import time
 
 import requests
 
-from promptflow._cli._pf._service import _start_background_service_on_unix, _start_background_service_on_windows
-from promptflow._sdk._service.utils.utils import get_pfs_port
-
 
 def invoke_prompt_flow_service() -> str:
     # invoke prompt flow service as a standby service
     # so use some private APIs, instead of existing API
     # then this port won't be recorded in pf.config
+    from promptflow._cli._pf._service import _start_background_service_on_unix, _start_background_service_on_windows
+    from promptflow._sdk._service.utils.utils import get_pfs_port
+
     port = str(get_pfs_port())
     if platform.system() == "Windows":
         _start_background_service_on_windows(port)
