+++ src/promptflow-recording/recordings/local/node_cache.shelve.dir
@@ -110,3 +110,4 @@
 '07226955e1e29e7dc59073f46aabbde06cb4d6cf', (613376, 11396)
 'f7ec26ad9b25d8c0391b005443cb2707d4212360', (625152, 2616)
 '5dbada5cafa016d2171e6c8643ba4c8397f03d7f', (628224, 2770)
+'ab3563e99d1ee052e067952ab332536a6bf5c025', (631296, 1674)
